Move *_at_point closer to where they're used
