+++ gaphas/segment.py
@@ -6,14 +6,13 @@ from cairo import ANTIALIAS_NONE
 from gi.repository import Gtk
 
 from gaphas.aspect.connector import ConnectionSink
-from gaphas.aspect.finder import item_at_point
 from gaphas.aspect.handlemove import HandleMove, ItemHandleMove
 from gaphas.canvas import Canvas
 from gaphas.connector import Handle, LinePort
 from gaphas.geometry import distance_line_point, distance_point_point_fast
 from gaphas.item import Line, matrix_i2i
 from gaphas.solver import WEAK
-from gaphas.tool.itemtool import MoveType
+from gaphas.tool.itemtool import MoveType, item_at_point
 from gaphas.view import Selection
