+++ tests/test_tool_item.py
@@ -1,6 +1,15 @@
+import math
+
 from gi.repository import Gtk
 
-from gaphas.tool.itemtool import DragState, item_tool, on_drag_begin
+from gaphas.item import Element as Box
+from gaphas.tool.itemtool import (
+    DragState,
+    handle_at_point,
+    item_at_point,
+    item_tool,
+    on_drag_begin,
+)
 
 
 class MockEvent:
@@ -52,3 +61,47 @@ def test_start_move_handle_on_click(view, box, window):
     assert drag_state.moving
     assert next(iter(drag_state.moving)).item is box
     assert next(iter(drag_state.moving)).handle is box.handles()[0]
+
+
+def test_get_item_at_point(view, box):
+    """Hover tool only reacts on motion-notify events."""
+    box.width = 50
+    box.height = 50
+
+    assert item_at_point(view, (10, 10)) is box
+    assert item_at_point(view, (60, 10)) is None
+
+
+def test_get_unselected_item_at_point(view, box):
+    box.width = 50
+    box.height = 50
+    view.selection.select_items(box)
+
+    assert item_at_point(view, (10, 10)) is box
+    assert item_at_point(view, (10, 10), selected=False) is None
+
+
+def test_get_handle_at_point(view, canvas, connections):
+    box = Box(connections)
+    box.min_width = 20
+    box.min_height = 30
+    box.matrix.translate(20, 20)
+    box.matrix.rotate(math.pi / 1.5)
+    canvas.add(box)
+
+    i, h = handle_at_point(view, (20, 20))
+    assert i is box
+    assert h is box.handles()[0]
+
+
+def test_get_handle_at_point_at_pi_div_2(view, canvas, connections):
+    box = Box(connections)
+    box.min_width = 20
+    box.min_height = 30
+    box.matrix.translate(20, 20)
+    box.matrix.rotate(math.pi / 2)
+    canvas.add(box)
+
+    i, h = handle_at_point(view, (20, 20))
+    assert i is box
+    assert h is box.handles()[0]
